chore: Clean up for sglang multimodal example (#3420)
