+++ components/backends/sglang/launch/agg_router.sh
@@ -12,7 +12,7 @@ cleanup() {
 trap cleanup EXIT INT TERM
 
 # run clear_namespace
-python3 -m dynamo.sglang.utils.clear_namespace --namespace dynamo
+python3 -m dynamo.sglang.clear_namespace --namespace dynamo
 
 # run ingress
 python -m dynamo.frontend --router-mode kv --http-port=8000 &
