+++ components/backends/sglang/launch/multimodal_disagg.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 # SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-set -e
-trap 'echo Cleaning up...; kill 0' EXIT
+
+# Setup cleanup trap
+cleanup() {
+    echo "Cleaning up background processes..."
+    kill $DYNAMO_PID $PREFILL_PID 2>/dev/null || true
+    wait $DYNAMO_PID $PREFILL_PID 2>/dev/null || true
+    echo "Cleanup complete."
+}
+trap cleanup EXIT INT TERM
 
 # Default values
 MODEL_NAME="Qwen/Qwen2.5-VL-7B-Instruct"
@@ -45,8 +52,12 @@ fi
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 SGLANG_BACKEND_DIR="$SCRIPT_DIR/src"
 
+# run clear_namespace
+python3 -m dynamo.sglang.clear_namespace --namespace dynamo
+
 # run ingress
-python -m dynamo.frontend --http-port=8000 &
+python3 -m dynamo.frontend --http-port=8000 &
+DYNAMO_PID=$!
 
 # run SGLang multimodal processor
 python3 -m dynamo.sglang --multimodal-processor --model-path "$MODEL_NAME" --chat-template "$CHAT_TEMPLATE" &
@@ -55,6 +66,8 @@ python3 -m dynamo.sglang --multimodal-processor --model-path "$MODEL_NAME" --cha
 CUDA_VISIBLE_DEVICES=0 python3 -m dynamo.sglang --multimodal-encode-worker --model-path "$MODEL_NAME" --chat-template "$CHAT_TEMPLATE" &
 
 # run SGLang multimodal prefill worker
+# TODO: Remove disable-radix-cache once the issue is fixed.
+# See https://github.com/sgl-project/sglang/pull/11203.
 CUDA_VISIBLE_DEVICES=1 python3 -m dynamo.sglang \
   --multimodal-worker \
   --model-path "$MODEL_NAME" \
@@ -65,6 +78,7 @@ CUDA_VISIBLE_DEVICES=1 python3 -m dynamo.sglang \
   --disaggregation-mode prefill \
   --disaggregation-bootstrap-port 12345 \
   --host 0.0.0.0 \
+  --disable-radix-cache \
   --disaggregation-transfer-backend nixl &
 
 # run SGLang multimodal decode worker
