+++ components/src/dynamo/sglang/multimodal_utils/__init__.py
@@ -1,0 +1,18 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from dynamo.sglang.multimodal_utils.multimodal_chat_processor import (
+    multimodal_request_to_sglang,
+    process_sglang_stream_response,
+)
+from dynamo.sglang.multimodal_utils.multimodal_encode_utils import (
+    encode_image_embeddings,
+)
+from dynamo.sglang.multimodal_utils.multimodal_image_loader import ImageLoader
+
+__all__ = [
+    "multimodal_request_to_sglang",
+    "process_sglang_stream_response",
+    "encode_image_embeddings",
+    "ImageLoader",
+]
