+++ components/src/dynamo/sglang/multimodal_utils/multimodal_image_loader.py
